hotfix/filipe_soua/cartstore | Ajustes de pastas e codigos
- Pastas realocadas para o modulo cart
- Modulos add card e form_payment tambem foram realocados para o modulo
  cart
- Remoção de modelos de itens e lista não necessários
- Novo modelo new_card para o cartão
- Refatoramento do codigos relacionados a sidebar
