+++ lib/app/global/models/entities/new_card.dart
@@ -1,0 +1,17 @@
+class NewCard {
+  final String? fourLastNumbers;
+  final String? brand;
+  final String? holder;
+  final String? valid;
+  final String? cpf;
+  final String? nickname;
+
+  NewCard({
+    this.fourLastNumbers,
+    this.brand,
+    this.holder,
+    this.valid,
+    this.cpf,
+    this.nickname,
+  });
+}
