+++ lib/app/modules/cart/cart_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_modular/flutter_modular.dart';
-import 'package:pscomidas/app/global/widgets/custom_page/custom_page.dart';
 import 'package:pscomidas/app/modules/cart/cart_store.dart';
 import 'package:flutter/material.dart';
+import 'package:pscomidas/app/modules/cart/components/cart_drawer/cart_drawer.dart';
 
 class CartPage extends StatefulWidget {
   final String title;
@@ -15,6 +15,6 @@ class CartPageState extends State<CartPage> {
 
   @override
   Widget build(BuildContext context) {
-    return CustomPage(body: Container());
+    return CartDrawer(placeHolder: store.placeholder);
   }
 }
